requirements.txt
Flask
psycopg2-binary
python-dotenv
boto3
# otras dependencias
